Refactored search bar UI, separated to subcomponents, separated controller and view.
