+++ App.js
@@ -2,7 +2,7 @@ import 'react-native-gesture-handler';
 import { NavigationContainer } from '@react-navigation/native';
 import { StatusBar } from 'expo-status-bar';
 import React from 'react';
-import SearchUserScreen from './components/search/SearchUserScreen';
+import SearchUserScreenController from './components/search/SearchUserScreenController';
 import { StyleSheet, Text, View, Button } from 'react-native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
@@ -12,7 +12,7 @@ export default function App() {
   // Temporary to test navigation
   const Home = ({ navigation }) => (
     <View style={styles.container} >
-      <Text>Vinson created this repository.</Text>
+      <Text style={{ fontSize: 24, textAlign: "center", margin: 8 }}>Temporary Home Menu</Text>
       <Button onPress={() => navigation.navigate("Page")} title="Other Page" />
       <Button onPress={() => navigation.navigate("Search")} title="Search Page" />
     </View >
@@ -29,13 +29,19 @@ export default function App() {
       <Stack.Navigator
         screenOptions={{
           headerStyle: {
-            backgroundColor: "orange",
-          }
+            backgroundColor: "grey"
+          },
+          headerTitleStyle: {
+            color: "#fff"
+          },
+          headerTintColor: "#fff",
+          headerShadowVisible: false,
+          animation: "slide_from_left"
         }}
       >
         <Stack.Screen component={Home} name="Home" />
         <Stack.Screen component={Page} name="Page" />
-        <Stack.Screen component={SearchUserScreen} name="Search" />
+        <Stack.Screen component={SearchUserScreenController} name="Search" />
       </Stack.Navigator>
       <StatusBar style="light" />
     </NavigationContainer>
